requirements.txt
openai>=1.30.0
pyairtable>=2.3.0
pydantic>=2.6.0
jsonschema>=4.22.0
python-docx>=1.1.0
PyMuPDF>=1.24.0
pdfplumber>=0.11.0
python-dotenv>=1.0.0
redis>=5.0.0
tenacity>=8.3.0
